+++ src/test/java/com/palominolabs/metrics/guice/MyException.java
@@ -1,6 +1,6 @@
 package com.palominolabs.metrics.guice;
 
-public class MyException extends RuntimeException {
+class MyException extends RuntimeException {
 
     private static final long serialVersionUID = 1L;
 }